+++ src/main/java/roomescape/service/ReservationService.java
@@ -1,13 +1,44 @@
 package roomescape.service;
 
 import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationDao;
+import roomescape.dao.ReservationTimeDao;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationRequest;
 import roomescape.dto.ReservationResponse;
+import roomescape.idgenerator.AutoIncrementIdGenerator;
+import roomescape.idgenerator.IdGenerator;
 
-public interface ReservationService {
-    List<ReservationResponse> findAll();
+@Service
+public class ReservationService {
+    private final ReservationDao reservationDao;
+    private final ReservationTimeDao reservationTimeDao;
+    private final IdGenerator idGenerator = new AutoIncrementIdGenerator();
 
-    ReservationResponse save(ReservationRequest reservationRequest);
+    public ReservationService(ReservationDao reservationDao,
+                              ReservationTimeDao reservationTimeDao) {
+        this.reservationDao = reservationDao;
+        this.reservationTimeDao = reservationTimeDao;
+    }
 
-    boolean deleteById(long id);
+    public List<ReservationResponse> findAll() {
+        List<Reservation> reservations = reservationDao.findAll();
+        return reservations.stream()
+                .map(ReservationResponse::new)
+                .toList();
+    }
+
+    public ReservationResponse save(ReservationRequest reservationRequest) {
+        long timeId = reservationRequest.timeId();
+        ReservationTime time = reservationTimeDao.findById(timeId);
+        Reservation reservation = reservationRequest.toDomain(idGenerator.generateNewId(), time);
+        reservationDao.save(reservation);
+        return new ReservationResponse(reservation);
+    }
+
+    public boolean deleteById(long id) {
+        return reservationDao.deleteById(id);
+    }
 }
