+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,13 +1,38 @@
 package roomescape.service;
 
 import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.dao.ReservationTimeDao;
+import roomescape.domain.ReservationTime;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
+import roomescape.idgenerator.AutoIncrementIdGenerator;
+import roomescape.idgenerator.IdGenerator;
 
-public interface ReservationTimeService {
-    List<ReservationTimeResponse> findAll();
+@Service
+public class ReservationTimeService {
+    private final ReservationTimeDao reservationTimeDao;
+    private final IdGenerator idGenerator = new AutoIncrementIdGenerator();
 
-    ReservationTimeResponse save(ReservationTimeRequest reservationTimeRequest);
+    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+        this.reservationTimeDao = reservationTimeDao;
+    }
 
-    boolean deleteById(long id);
+    public List<ReservationTimeResponse> findAll() {
+        return reservationTimeDao.findAll()
+                .stream()
+                .map(ReservationTimeResponse::new)
+                .toList();
+    }
+
+    public ReservationTimeResponse save(ReservationTimeRequest reservationTimeRequest) {
+        ReservationTime reservationTime = reservationTimeRequest.toDomain(
+                idGenerator.generateNewId());
+        reservationTimeDao.save(reservationTime);
+        return new ReservationTimeResponse(reservationTime);
+    }
+
+    public boolean deleteById(long id) {
+        return reservationTimeDao.deleteById(id);
+    }
 }
